requirements.txt
lxml
tqdm
pillow
future
git+https://github.com/mon/kbinxml.git